03-02-2025 - Shared frontend code
