+++ src/frontend/src/styles/variables.scss
@@ -1,0 +1,18 @@
+// Brand colors
+$primary: #4C64E2;
+$primary-light: #F0F3FF;
+$primary-dark: #3b4fd9;
+
+// Text colors
+$text-primary: #333333;
+$text-secondary: #666666;
+$text-muted: #5d6b8a;
+
+// Backgrounds
+$bg-gradient: linear-gradient(180deg, #fafbff 0%, #f8f9ff 100%);
+
+// Layout
+$border-radius: 16px;
+$max-width: 1280px;
+$shadow-light: 0 2px 20px rgba($primary, 0.08), 0 1px 8px rgba($primary, 0.05);
+$shadow-medium: 0 6px 30px rgba($primary, 0.08);
